+++ llama_index/query_engine/citation_query_engine.py
@@ -266,7 +266,7 @@ class CitationQueryEngine(BaseQueryEngine):
         )
 
         retrieve_id = self.callback_manager.on_event_start(CBEventType.RETRIEVE)
-        nodes = self._retriever.retrieve(query_bundle)
+        nodes = self.retrieve(query_bundle)
         nodes = self._create_citation_nodes(nodes)
         self.callback_manager.on_event_end(
             CBEventType.RETRIEVE,
@@ -293,7 +293,7 @@ class CitationQueryEngine(BaseQueryEngine):
         )
 
         retrieve_id = self.callback_manager.on_event_start(CBEventType.RETRIEVE)
-        nodes = self._retriever.retrieve(query_bundle)
+        nodes = self.retrieve(query_bundle)
         nodes = self._create_citation_nodes(nodes)
         self.callback_manager.on_event_end(
             CBEventType.RETRIEVE,